ref(slack): simplify input ownership
